+++ backend/requirements-minimal.txt
@@ -1,0 +1,47 @@
+# Minimal production requirements for Docker deployment
+# Core dependencies only to avoid conflicts
+
+# Web framework
+fastapi==0.116.1
+uvicorn==0.35.0
+starlette==0.47.2
+
+# Database
+asyncpg==0.30.0
+psycopg2==2.9.10
+SQLAlchemy==2.0.43
+alembic==1.16.4
+
+# Authentication and security
+python-jose==3.5.0
+python-multipart==0.0.20
+passlib==1.7.4
+bcrypt==4.3.0
+PyJWT==2.10.1
+
+# HTTP and async
+httpx==0.28.1
+aiohttp==3.12.15
+requests==2.32.4
+
+# Data processing
+pydantic==2.11.7
+pydantic-settings==2.10.1
+python-dotenv==1.1.1
+
+# Utilities
+click==8.2.1
+rich==14.1.0
+tqdm==4.67.1
+
+# File handling
+aiofiles==24.1.0
+Pillow==11.3.0
+PyMuPDF==1.25.3
+
+# AI/ML (minimal set)
+numpy==2.3.2
+scipy==1.16.1
+
+# Google AI (current unified SDK)
+google-genai>=0.3.0
