+++ backend/requirements-prod.txt
@@ -45,14 +45,13 @@ filelock==3.19.1
 filetype==1.2.0
 frozenlist==1.7.0
 fsspec==2025.7.0
-google-ai-generativelanguage>=0.6.18,<0.7.0
+# Google AI packages - using current unified SDK
+google-genai>=0.3.0
 google-api-core==2.25.1
 google-api-python-client==2.179.0
 google-auth==2.40.3
 google-auth-httplib2==0.2.0
 google-auth-oauthlib==1.2.2
-google-genai
-google-generativeai
 googleapis-common-protos==1.70.0
 greenlet==3.2.4
 groq==0.31.0
@@ -81,7 +80,7 @@ jsonpointer==3.0.0
 jsonschema==4.25.0
 jsonschema-specifications==2025.4.1
 langchain-core==0.3.74
-langchain-google-genai
+langchain-google-genai>=2.2.0
 langchain-mcp-adapters==0.1.9
 langgraph==0.6.5
 langgraph-checkpoint==2.1.1
