+++ frontend/next.config.mjs
@@ -1,39 +1,1 @@
-import { fileURLToPath } from 'url'
-import { dirname, resolve } from 'path'
-
-const __filename = fileURLToPath(import.meta.url)
-const __dirname = dirname(__filename)
-
-/** @type {import('next.js').NextConfig} */
-const nextConfig = {
-  eslint: {
-    ignoreDuringBuilds: true,
-  },
-  typescript: {
-    ignoreBuildErrors: true,
-  },
-  images: {
-    unoptimized: true,
-  },
-  experimental: {
-    esmExternals: true,
-  },
-  webpack: (config, { isServer }) => {
-    // Ensure proper path resolution
-    config.resolve.alias = {
-      ...config.resolve.alias,
-      '@': resolve(__dirname),
-    }
-    return config
-  },
-  async rewrites() {
-    return [
-      {
-        source: '/api/v1/:path*',
-        destination: 'http://localhost:8000/api/v1/:path*',
-      },
-    ]
-  },
-}
-
-export default nextConfig
+
